Add variable animation speed and mech sprite

+++ src/systems/update_sprite_index.rs
@@ -3,7 +3,7 @@ use bevy::prelude::*;
 
 pub fn update_sprite_index(mut sprites: Query<(&mut TextureAtlasSprite, &SpriteAnimation)>) {
     sprites.for_each_mut(|(mut sprite, animation)| {
-        sprite.index =
-            (animation.current_angle * animation.num_frames_per_angle) + animation.current_frame;
+        sprite.index = (animation.current_angle * animation.num_frames_per_angle)
+            + animation.current_frame as usize;
     })
 }
